Struct claude::Currency

pub struct Currency {
    pub symbol: Option<Symbol>,
    pub value: i64,
}

Represents currency through an optional symbol and amount of coin.

Each 100 coins results in a banknote. (100 is formatted as 1.00) The currency will be formatted as such: Currency(Some('$'), 432) ==> "$4.32"

Fields
symbol: Option<Symbol>

Currency symbol

pick any of €, £, $, ¥ etc...

value: i64

value in the smallest possible unit

Methods

impl Currency

pub fn new() -> Currency

Creates a blank Currency as Currency(None, 0)

Examples

let mut c = Currency::new();

pub fn from_value(value: i64) -> Currency

Initialize from i64

pub fn postfix(&self) -> Postfix

Returns an object that implements Display for different methods of printing currency.

pub fn prefix(&self) -> Prefix

Returns an object that implements Display for different methods of printing currency.

pub fn as_float(&self) -> f64

Returns the value as float

Warning, do not use this for calculation, this is for displaying only!

pub fn value(&self) -> i64

Returns the inner value

pub fn symbol(&self) -> Option<Symbol>

Returns the inner symbol