Struct claude::Currency

pub struct Currency(pub Option<char>, pub i64);

Represents currency through an optional symbol and amount of coin.

Each 100 coins results in a banknote. (100 is formatted as 1.00) The currency will be formatted as such: Currency(Some('$'), 432) ==> "$4.32"

Methods

impl Currency

fn new() -> Currency

Creates a blank Currency as Currency(None, 0)

Examples

use claude::Currency;

let mut c = Currency::new();

fn postfix(&self) -> Postfix

Returns an object that implements Display for different methods of printing currency.

fn prefix(&self) -> Prefix

Returns an object that implements Display for different methods of printing currency.

fn as_float(&self) -> f64

Returns the value as float

Warning, do not use this for calculation, this is for displaying only!

fn value(&self) -> i64

Returns the inner value